Move LibKstatFunctions out of the util tree to oshi.ffm.platform.unix.solaris
`LibKstatFunctions` extends `ForeignFunctions` — it's the libkstat FFM
mapping, not a utility class. It was sitting in
`oshi.ffm.util.platform.unix.solaris.*` next to `KstatUtilFFM` (a real
utility) and `Kstat2Functions` (a probe-only stub). Move it to
`oshi.ffm.platform.unix.solaris.*` alongside the other libc mappings.

`Kstat2Functions` retains its same-package `{@link LibKstatFunctions}`
javadoc reference, now resolved through an explicit import.
