+++ oshi-core-ffm/src/main/java/oshi/ffm/platform/unix/solaris/LibKstatFunctions.java
@@ -2,7 +2,7 @@
  * Copyright 2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.ffm.util.platform.unix.solaris;
+package oshi.ffm.platform.unix.solaris;
 
 import static java.lang.foreign.ValueLayout.ADDRESS;
 import static java.lang.foreign.ValueLayout.JAVA_BYTE;
