+++ oshi-core-ffm/src/main/java/oshi/hardware/platform/unix/solaris/SolarisHWDiskStoreFFM.java
@@ -17,9 +17,9 @@ import oshi.annotation.concurrent.ThreadSafe;
 import oshi.driver.common.unix.solaris.disk.Iostat;
 import oshi.driver.common.unix.solaris.disk.Lshal;
 import oshi.driver.common.unix.solaris.disk.Prtvtoc;
+import oshi.ffm.platform.unix.solaris.LibKstatFunctions;
 import oshi.ffm.util.platform.unix.solaris.KstatUtilFFM;
 import oshi.ffm.util.platform.unix.solaris.KstatUtilFFM.KstatChain;
-import oshi.ffm.util.platform.unix.solaris.LibKstatFunctions;
 import oshi.hardware.HWDiskStore;
 import oshi.hardware.HWPartition;
 import oshi.hardware.common.AbstractHWDiskStore;
